requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cloudinary==1.29.0
Django==4.0.4
djangorestframework==3.13.1
psycopg2-binary==2.9.3
python-decouple==3.6
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
